fix(refactor): type 정리
