+++ src/shared/types/index.ts
@@ -1,4 +1,5 @@
 import { z } from 'zod/v4-mini'
+import type { CreateNextContextOptions } from '@trpc/server/adapters/next'
 
 export type ZodType<T extends z.ZodMiniType> = z.infer<T>
 
@@ -13,3 +14,11 @@ export interface PrismaPaginationOptions {
 	skip?: number
 	take?: number
 }
+
+interface CreateAppContextOptions {
+	req: Request
+}
+
+export type CreateContextOptions =
+	| CreateNextContextOptions
+	| CreateAppContextOptions
