--- src/types/error-modal.ts
@@ -1,6 +1,0 @@
-export interface ErrorModalProps {
-	open: boolean
-	onClose: () => void
-	title?: string
-	message: string
-}
