+++ src/types/icon.ts
@@ -1,0 +1,7 @@
+export interface KakaoTalkIconProps {
+	size?: number
+}
+
+export interface NaverCafeIconProps {
+	size?: number
+}
