+++ src/types/login.ts
@@ -1,0 +1,4 @@
+export interface LoginModalProps {
+	open: boolean
+	onClose: () => void
+}
